requirements.txt
CherryPy==3.5.0
asyncio==3.4.1
gevent==1.0.1
gevent-websocket==0.9.3
greenlet==0.4.3
mock==1.0.1
pyserial==2.7
ws4py==0.3.4
wsgiref==0.1.2
